Update layer folder structure
