--- src/components/Map/Layers/index.ts
@@ -1,40 +1,0 @@
-// Exports the layer configurations to the Map component.
-
-import buildingEnergyCertificates from './Buildings/BuildingEnergyCertificates'
-// import finlandBuildings from './Buildings/finlandBuildings'
-import hsySolarpotential from './Buildings/HsySolarPotential'
-import helsinkiBuildings from './Buildings/HelsinkiBuildings'
-import snowCoverLoss from './SnowCoverLoss'
-import airQuality from './AirQuality'
-import finlandBogs from './Wetlands/FinlandBogs'
-import ciforPeatlands from './Wetlands/CiforPeatlands'
-import ciforWetlands from './Wetlands/CiforWetlands'
-import metsaanEteBasic from './Biodiversity/MetsaanEteBasic'
-import metsaanEteImportant from './Biodiversity/MetsaanEteImportant'
-import zonation from './Biodiversity/Zonation'
-import natura2000 from './Biodiversity/Natura2000'
-import hansen from './Forests/Hansen'
-import finlandMatureForests from './Forests/FinlandMatureForests'
-import mangroveForests from './Forests/MangroveForests'
-import treePlantations from './Forests/TreePlantations'
-import terramonitor from './Background/Terramonitor'
-
-export const layerConfs = [
-  buildingEnergyCertificates,
-  snowCoverLoss,
-  airQuality,
-  hsySolarpotential,
-  helsinkiBuildings,
-  finlandBogs,
-  ciforPeatlands,
-  ciforWetlands,
-  metsaanEteBasic,
-  metsaanEteImportant,
-  zonation,
-  natura2000,
-  hansen,
-  finlandMatureForests,
-  mangroveForests,
-  treePlantations,
-  terramonitor,
-]
