+++ src/components/Map/layers/main/Buildings/HsySolarPotential/index.ts
