+++ src/components/Map/layers/main/Buildings/HsySolarPotential/layerConf.ts
