+++ src/components/Map/layers/main/SnowCoverLoss/index.ts
