+++ src/components/Map/layers/main/SnowCoverLoss/layerConf.ts
