+++ src/components/Map/layers/main/Wetlands/FinlandBogs/layerConf.ts
